requirements.txt
click==6.7
Django==1.11.2
Flask==0.12.2
image==1.5.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
Pillow==4.1.1
pytz==2017.2
Werkzeug==0.12.2
